transfer origin and proxy handling
Add `data->state.origin` as the origin the transfer
is sending the current request to/gets the response
from. Use it for request specific properties like
authentication, hsts and cookie handling, etc.

Unless talking to a forwarding HTTP proxy (e.g.
not tunneling), `data->state.origin` and `conn->origin`
are the same.

With a forwarding HTTP proxy in play, `conn->origin`
is set to `conn->http_proxy.peer` and
`conn->bits.origin_is_proxy` (a new bit) is set.

Remove the connection bits, now replaced with:

* `conn->bits.socksproxy` -> `conn->socks_proy.peer`
* `conn->bits.httpproxy` -> `conn->http_proy.peer`
* `conn->bits.proxy` -> `(conn->socks_proy.peer || conn->http_proy.peer`)
* `conn->bits.tunnel_proxy` -> (`conn->http_proy.peer && !conn->bits.origin_is_proxy`)
* `(conn->bits.httpproxy && !conn->bits.tunnel_proxy)` -> `conn->bits.origin_is_proxy`

Rename `noproxy.[ch]` to `proxy.[ch]`. Move the connection
proxy setup code from `url.c` to `proxy.c`.

Remove `data->info.conn_remote_port` as no one uses it.

Add test_40_02b for a SOCKS connection to a forwarding
HTTPS proxy.

Update internal documentation about peers and creds.
