+++ lib/transfer.c
@@ -493,6 +493,7 @@ CURLcode Curl_pretransfer(struct Curl_easy *data)
   /* initial transfer request coming up, forget the initial origin
    * from a previous perform() on this handle. */
   Curl_peer_unlink(&data->state.initial_origin);
+  Curl_peer_unlink(&data->state.origin);
   data->state.requests = 0;
   data->state.followlocation = 0; /* reset the location-follow counter */
   data->state.this_is_a_follow = FALSE; /* reset this */
